pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contexthub-amc"
version = "0.1.0"
description = "ContextHub Agent Memory Core (AMC) — trajectory commit/retrieve, graph + vector stores"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "ContextHub" }]

# 运行时依赖：后续新增库请同步改这里，并在 README「依赖变更记录」中记一笔。
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "neo4j>=5.0",
    "chromadb>=0.5",
    "openai>=1.0",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.0",
]

[project.scripts]
amc-commit-trajectory = "cli.commit_trajectory:main"
amc-check-llm = "cli.check_llm_connectivity:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["api*", "app*", "core*", "domain*", "infra*", "cli*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
pythonpath = ["src"]
markers = [
    "unit: fast unit tests, no external services",
    "integration: needs Neo4j/Chroma/network",
    "m1: milestone M1 acceptance cases",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
